% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Arab.Spring}
\alias{Arab.Spring}
\title{Determinants of the Arab Spring Uprising}
\format{
A data frame with 20 rows and 10 variables, measured as fuzzy-set
membership scores:
\describe{
\item{Gdppc}{Per-capita GDP.}
\item{Gini}{Economic inequality.}
\item{Unemp}{Unemployment.}
\item{Urban}{Degree of urbanization.}
\item{Youth}{Size of youth bulge.}
\item{Mobile}{Mobile phone usage.}
\item{Internet}{Internet penetration.}
\item{Fuel}{Fuel-dependence of economy.}
\item{Pol}{Regime fragility.}
\item{Success}{Social movement success.}
}
}
\source{
Hussain, Muzammil M., and Philip N. Howard. 2013. "What Best Explains
Successful Protest Cascades? ICTs and the Fuzzy Causes of the Arab Spring."
\emph{International Studies Review} 15(1): 48--66.
}
\usage{
Arab.Spring
}
\value{
A data.frame object
}
\description{
A dataset containing fuzzy-set membership scores for ten sets of state
characteristics across 20 states in northern Africa and the Arab Peninsula
at the time of the Arab Spring uprisings of 2010--12.
}
\keyword{datasets}
