% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitPage.R
\name{fitPage}
\alias{fitPage}
\title{Shortcut for \code{flextable} to fit the MS Word output well.}
\usage{
fitPage(ft, pgwidth = 6)
}
\arguments{
\item{ft}{A \code{\link[flextable]{flextable}} object.}

\item{pgwidth}{The width of the table in the Word document; the default is 6.}
}
\description{
Shortcut for \code{flextable} to fit the MS Word output well.
}
