% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_goldenBlack.R
\name{gb_cols}
\alias{gb_cols}
\title{Function to extract colors as hex codes.}
\usage{
gb_cols(...)
}
\arguments{
\item{...}{A character vector indicating the names of colors. Options includes "gold", "black", "orange", "red", "light gree", "green", "light blue", "blue", "light grey", "grey", and "dark grey".}
}
\value{
A character of hex codes.
}
\description{
Function to extract colors as hex codes.
}
