% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_goldenBlack.R
\name{gb_pal}
\alias{gb_pal}
\title{Function to interpolate a gold & black color palette.}
\usage{
gb_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{A character vector indicating the name of palette in gb_palettes. Available palettes:
\itemize{
\item \code{main}: Gold and black colors.
\item \code{tricol}: Gold, black, and dark grey to create a gradual effect.
\item \code{digitMixed}: Five-pack colors specified for digital publications.
\item \code{printMixed}: Five-pack colors specified for printed publications.
\item \code{full}: A palette including all the colors \code{gb_cols} can call.
}}

\item{reverse}{A logic vector indicating whether the palette should be reversed; the default is FALSE.}

\item{...}{Additional arguments to pass to \code{colorRampPalette()}}
}
\value{
A function that takes an integer argument (the required number of colors) and returns a character vector of colors interpolating the given sequence.
}
\description{
Function to interpolate a gold & black color palette.
}
