% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_maps_and_animation.R
\name{generate_maps}
\alias{generate_maps}
\title{Generate maps and animations for specified years}
\usage{
generate_maps(
  data_file,
  years,
  output_dir = "output",
  map_center = c(35.8617, 104.1954),
  zoom_level = 4,
  palette = "main",
  reverse_palette = TRUE,
  marker_radius = 1,
  legend_opacity = 0.7,
  width = 800,
  height = 900
)
}
\arguments{
\item{data_file}{A string, the path to the CSV file containing the data. The file should contain the following columns: \code{g_lat}, \code{g_lon}, \code{g_pro}, \code{g_city}, \code{type}, \code{year_set}.}

\item{years}{A numeric vector, a vector containing the years for which maps should be generated.}

\item{output_dir}{A string, optional, the output directory where the generated map images and animated GIF will be stored. Default is "output".}

\item{map_center}{A numeric vector, optional, the center point of the map, given as latitude and longitude coordinates. Default is \code{c(35.8617, 104.1954)} (center of China).}

\item{zoom_level}{A numeric value, optional, the zoom level of the map. Default is \code{4}.}

\item{palette}{A string, optional, the color palette for the map. Default is "main".}

\item{reverse_palette}{A logical, optional, whether to reverse the color palette. Default is TRUE.}

\item{marker_radius}{A numeric value, optional, the radius of the map markers. Default is \code{1}.}

\item{legend_opacity}{A numeric value, optional, the opacity of the legend. Default is \code{0.7}.}

\item{width}{A numeric value, optional, the width of the output map images. Default is \code{800}.}

\item{height}{A numeric value, optional, the height of the output map images. Default is \code{900}.}

\item{lat_col}{A string, optional, the column name for latitude. Default is \code{g_lat}.}

\item{lon_col}{A string, optional, the column name for longitude. Default is \code{g_lon}.}

\item{pro_col}{A string, optional, the column name for province. Default is \code{g_pro}.}

\item{city_col}{A string, optional, the column name for city. Default is \code{g_city}.}

\item{type_col}{A string, optional, the column name for type. Default is \code{type}.}
}
\value{
The function saves the generated map images and animated GIF to the specified output directory.
}
\description{
This function generates maps for specified years and saves these maps as image files.
Finally, it combines the generated map images into an animated GIF.
}
\examples{
data example:

  wave                 name year_set    g_lat    g_lon  g_pro g_city type
1    7 中国共产党历史展览馆     2021 40.00379 116.3994 北京市 北京市    7
2    7 中央礼品文物管理中心     2021 39.89785 116.4109 北京市 北京市    7
3    7           中国美术馆     2021 39.92535 116.4090 北京市 北京市    8
4    7       中国电影博物馆     2021 39.99636 116.5206 北京市 北京市    8
5    7   中国邮政邮票博物馆     2021 39.91081 116.4311 北京市 北京市    8
6    7       中国钱币博物馆     2021 39.90215 116.3949 北京市 北京市    8

generate_maps(data_file = ".../file.csv", years = c(1997, 2001, 2005), output_dir = "output", palette = "digitMixed")

generate_animation(output_dir = "output", years = c(1997, 2001, 2005), gif_name = "my_animation.gif", fps = 1)

}
