% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodChinaMap.R
\name{goodChinaMap}
\alias{goodChinaMap}
\title{Generate a Customizable Map of China with mark specific Provinces and Cities}
\usage{
goodChinaMap(
  province_data = NULL,
  city_data = NULL,
  saveDir = "output",
  width = 1200,
  height = 800,
  province_color = "yellow",
  province_border_color = "blue",
  city_color = "red",
  city_border_color = "black",
  ...
)
}
\arguments{
\item{province_data}{A character vector containing the names of provinces to be highlighted on the map. Default is \code{NULL}.}

\item{city_data}{A character vector containing the names of cities to be marked on the map. Default is \code{NULL}.}

\item{saveDir}{A character string specifying the directory where the output PNG file will be saved. Default is \code{"output"}.}

\item{width}{The width of the saved PNG file, in pixels. Default is \code{1200}.}

\item{height}{The height of the saved PNG file, in pixels. Default is \code{800}.}

\item{province_color}{A character string specifying the fill color for the highlighted provinces. Default is \code{"yellow"}.}

\item{province_border_color}{A character string specifying the border color for the highlighted provinces. Default is \code{"blue"}.}

\item{city_color}{A character string specifying the color for the city markers. Default is \code{"red"}.}

\item{city_border_color}{A character string specifying the border color for the city markers. Default is \code{"black"}.}
}
\value{
A character string with the file path of the generated PNG map.
}
\description{
This function generates a map of China highlighting specified provinces and cities.
The colors of the provinces, province borders, cities, and city borders can be customized.
The generated map is saved as a PNG file in the specified directory.
}
\examples{

  goodChinaMap(
  province_data = c("吉林省", "山东省"), 
  city_data = c("北京", "海口"), 
  province_color = "grey", 
  province_border_color = "purple", 
  city_color = "black", 
  city_border_color = "gold"
  )
  
}
