% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{leafletcn.map.names}
\alias{leafletcn.map.names}
\title{A data file of words and frequency from tm package}
\format{
A data set cities and file names
}
\usage{
leafletcn.map.names
}
\description{
A data file of words and frequency from tm package
}
\keyword{datasets}
