% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_goldenBlack.R
\name{scale_fill_gb}
\alias{scale_fill_gb}
\title{Color scale constructors for gold & black colors}
\usage{
scale_fill_gb(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{A character vector indicating the name of palette in \code{gb_palettes}. Available palettes:
\itemize{
\item \code{main}: Gold and black colors.
\item \code{tricol}: Gold, black, and dark grey to create a gradual effect.
\item \code{digitMixed}: Five-pack colors specified for digital publications.
\item \code{printMixed}: Five-pack colors specified for printed publications.
\item \code{full}: A palette including all the colors \code{gb_cols} can call.
}}

\item{discrete}{A logic vector indicating whether color aesthetic is discrete or not; the default is "main".}

\item{reverse}{A logic vector indicating whether the palette should be reversed.}

\item{...}{Additional arguments passed to \code{discrete_scale()} or \code{scale_color_gradientn()}, used respectively when discrete is TRUE or FALSE}
}
\description{
Color scale constructors for gold & black colors
}
