% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{social.revolutions}
\alias{social.revolutions}
\title{Determinants of Social Revolutions}
\format{
A data frame with 20 rows and 3 variables, measured as degree of
fuzzy-set membership:
\describe{
\item{soc.rev}{Presence of social revolution.}
\item{breakdown}{Degree of state breakdown.}
\item{pop.ins}{Presence of popular insurrection.}
}
}
\source{
Ragin, Charles C. 2000. \emph{Fuzzy-Set Social Science}. Chicago:
University of Chicago Press, p. 220.
}
\usage{
social.revolutions
}
\value{
A data.frame object
}
\description{
A data frame containing hypothetical fuzzy-set membership scores for
three sets of state characteristics across twenty cases.
}
\keyword{datasets}
