% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.fsQCApt.R
\name{summary.fsQCApt}
\alias{summary.fsQCApt}
\title{Summarize Permutation Tests for fsQCA Data}
\usage{
\method{summary}{fsQCApt}(object, ...)
}
\arguments{
\item{object}{Object returned by \code{\link[=fsQCApermTest]{fsQCApermTest()}}.}

\item{...}{Additional parameters to pass on.}
}
\value{
Two matrices of values for counterexamples and consistency.
}
\description{
Displays observed values, confidence intervals, and raw and adjusted p-scores for both consistency and counterexamples following permutation test of fsQCA data.
}
\note{
Adopted from the archived CRAN package \pkg{QCAfalsePositive} by Bear
Braumoeller. The original package has been removed from CRAN and is no
longer maintained. It is included in \pkg{drhutools} for continued
accessibility.
}
\examples{
\donttest{
intersect  <- pmin(social.revolutions$breakdown, social.revolutions$pop.ins)
intersect2 <- pmin(social.revolutions$breakdown, (1 - social.revolutions$pop.ins))
intersect3 <- pmin((1 - social.revolutions$breakdown), social.revolutions$pop.ins)
intersect4 <- pmin((1 - social.revolutions$breakdown), (1 - social.revolutions$pop.ins))

test <- fsQCApermTest(y = social.revolutions$soc.rev,
  configs = list(BI = intersect, Bi = intersect2,
                 bI = intersect3, bi = intersect4),
  total.configs = 4)
summary(test)
}
}
\concept{permutation test fsQCA}
