% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textWrapper.R
\name{text_wrapper}
\alias{text_wrapper}
\title{Function to wrap}
\usage{
text_wrapper(label, dev_width = dev.size("in")[1], dev_scaler = 12)
}
\arguments{
\item{label}{A character string to wrap.}

\item{dev_width}{A numeric value specifying width of current device.}

\item{dev_scaler}{A numeric value to scale; the default is 12.}
}
\value{
A character vector with wrapped signs.
}
\description{
Function to wrap
}
\examples{
library(ggplot2)
ggplot(data = cars, aes(x = speed, y = dist)) +    
geom_smooth() +   
  labs(title = text_wrapper("A so very looooooooooooooooooooooooooooooong title"))
  
}
