% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_poly}
\alias{toy_poly}
\title{Toy Dataset for \code{goodmap}}
\format{
A geographic data frame with 6 rows and 5 variables:
\describe{
\item{id}{Optional data identifier.}
\item{city}{Prefectural names, required for city-level map plotting.}
\item{prov}{Provincial names, required for provincial-level map plotting.}
\item{animate_set}{A numeric vector used to group data for animated plotting.}
\item{value_set}{Values used to fill each polygon in the map.}
}
}
\usage{
toy_poly
}
\value{
A data.frame object
}
\description{
A sample dataset designed to illustrate the functionality of the \code{goodmap} function.
}
\keyword{datasets}
