% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folderSystem.R
\name{folderSystem}
\alias{folderSystem}
\title{Create Dr.Hu recommanded folder system for an academic paper project}
\usage{
folderSystem()
}
\value{
A set of folders under the working directory.
}
\description{
Create Dr.Hu recommanded folder system for an academic paper project
}
\details{
The system contains specific places to store codes, output, data, manuscript (referred images and submission versions), and miscellaneous documents. It helps researchers to organize materials and drafts during the whole process of academic writing from data collection to manuscript publication. This is the recommanded folder structure for cooperative projects with members of Dr.Hu's Amazing Team.
}
\examples{

# Not run

# folderSystem()

}
