% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{leafletcn.map.names}
\alias{leafletcn.map.names}
\title{Leaflet Map Names Dataset}
\format{
A data frame with city names and file names.
}
\usage{
leafletcn.map.names
}
\description{
A internal dataset providing city names and corresponding file names for mapping with Leaflet.
}
\keyword{datasets}
